Was able to fix the issue with params not matching
