+++ friends/src/components/friend.js
@@ -1,0 +1,15 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+export default props => {
+
+    return (
+        <div>
+
+            <Link to={`/friends/${props.id}`}>
+             {props.name}
+            </Link>
+
+        </div>
+    );
+};
