+++ friends/src/components/friendCard.js
@@ -1,0 +1,8 @@
+import React from "react";
+
+
+export default props => {
+    return (
+        <div> {props.name}</div>
+    );
+};
